[Cache] Addition to #5511 : execution should be at night between 0:00 and 4:59

+++ lib/Maintenance/Tasks/CacheCleanupTask.php
@@ -49,7 +49,8 @@ final class CacheCleanupTask implements TaskInterface
     public function execute()
     {
         $lockId = self::class;
-        if(!Lock::isLocked($lockId, 86400)) {
+        if(!Lock::isLocked($lockId, 86400) && date('H') <= 4) {
+            // execution should be only sometime between 0:00 and 4:59 -> less load expected
             Lock::lock($lockId);
             $this->logger->debug('Execute purge() on cache handler');
             $this->cacheHandler->purge();
